refactor: refactor file read methods
