+++ pages/blog/[slug].tsx
@@ -9,12 +9,14 @@ import {
   Divider,
   CSSReset,
 } from "@chakra-ui/react";
-import fs from "fs";
-import matter from "gray-matter";
 import md from "markdown-it";
 import { useMemo } from "react";
 import { PostData } from "../../features/blog/types";
 import { Prose } from "@nikolovlazar/chakra-ui-prose";
+import {
+  getFileReadStaticPaths,
+  getFileReadStaticProps,
+} from "../../features/file-read";
 
 export type BlogPostPageProps = { frontmatter: PostData; content: string };
 
@@ -71,19 +73,7 @@ export default function BlogPostPage(props: BlogPostPageProps) {
 
 // Generating the paths for each post
 export async function getStaticPaths() {
-  // Get list of all files from our posts directory
-  const files = fs.readdirSync("content/posts");
-  // Generate a path for each one
-  const paths = files.map((fileName) => ({
-    params: {
-      slug: fileName.replace(".md", ""),
-    },
-  }));
-  // return list of paths
-  return {
-    paths,
-    fallback: false,
-  };
+  return getFileReadStaticPaths("content/posts");
 }
 
 // Generate the static props for the page
@@ -92,12 +82,5 @@ export async function getStaticProps({
 }: {
   params: { slug: string };
 }) {
-  const fileName = fs.readFileSync(`content/posts/${slug}.md`, "utf-8");
-  const { data: frontmatter, content } = matter(fileName);
-  return {
-    props: {
-      frontmatter,
-      content,
-    },
-  };
+  return getFileReadStaticProps(slug);
 }
