+++ pages/events/[slug].tsx
@@ -9,21 +9,24 @@ import {
   Avatar,
   Divider,
   CSSReset,
-  Stack
+  Stack,
 } from "@chakra-ui/react";
-import fs from "fs";
-import matter from "gray-matter";
 import md from "markdown-it";
 import { useMemo } from "react";
 import { EventData } from "../../features/events/types";
 import { Prose } from "@nikolovlazar/chakra-ui-prose";
 import { useRouter } from "next/router";
+import {
+  getFileReadStaticPaths,
+  getFileReadStaticProps,
+} from "../../features/file-read";
 
 export type EventPostPageProps = { frontmatter: EventData; content: string };
 
 export default function BlogPostPage(props: EventPostPageProps) {
   const router = useRouter();
-  const { id, date, title, description, tags, location, url } = props.frontmatter;
+  const { id, date, title, description, tags, location, url } =
+    props.frontmatter;
   const tagsAsArray = useMemo(() => {
     if (!tags) return [];
     return tags.split(", ");
@@ -38,36 +41,38 @@ export default function BlogPostPage(props: EventPostPageProps) {
         background={primary[800]}
         rounded="xl"
       >
-        <Stack direction="row">
-
-        </Stack>
+        <Stack direction="row"></Stack>
         <Stack>
-        {title && (
-          <Heading size="lg" color={primary[50]}>
-            {title}
-          </Heading>
-        )}
-        {description && <Text>{description}</Text>}
-        {tags && (
-          <Box mt={3}>
-            {tagsAsArray.map((tag, i) => (
-              <Tag key={i} mr={2} mb={2}>
-                {tag}
-              </Tag>
-            ))}
-          </Box>
-        )}
-        {location && date && (
-          <Box mt={3} display="flex" alignItems="center" gap={3}>
-            <Text>
-              At <b>{location}</b> on <b>{date}</b>
-            </Text>
-          </Box>
-        )}
+          {title && (
+            <Heading size="lg" color={primary[50]}>
+              {title}
+            </Heading>
+          )}
+          {description && <Text>{description}</Text>}
+          {tags && (
+            <Box mt={3}>
+              {tagsAsArray.map((tag, i) => (
+                <Tag key={i} mr={2} mb={2}>
+                  {tag}
+                </Tag>
+              ))}
+            </Box>
+          )}
+          {location && date && (
+            <Box mt={3} display="flex" alignItems="center" gap={3}>
+              <Text>
+                At <b>{location}</b> on <b>{date}</b>
+              </Text>
+            </Box>
+          )}
         </Stack>
-        {
-          url && (<Box padding={4}><Button width="100%" onClick={() => router.push(url)}>Register Here</Button></Box>)
-        }
+        {url && (
+          <Box padding={4}>
+            <Button width="100%" onClick={() => router.push(url)}>
+              Register Here
+            </Button>
+          </Box>
+        )}
 
         <Divider mt={3} />
         <Prose>
@@ -83,19 +88,7 @@ export default function BlogPostPage(props: EventPostPageProps) {
 
 // Generating the paths for each post
 export async function getStaticPaths() {
-  // Get list of all files from our events directory
-  const files = fs.readdirSync("content/events");
-  // Generate a path for each one
-  const paths = files.map((fileName) => ({
-    params: {
-      slug: fileName.replace(".md", ""),
-    },
-  }));
-  // return list of paths
-  return {
-    paths,
-    fallback: false,
-  };
+  return getFileReadStaticPaths("content/events");
 }
 
 // Generate the static props for the page
@@ -104,12 +97,5 @@ export async function getStaticProps({
 }: {
   params: { slug: string };
 }) {
-  const fileName = fs.readFileSync(`content/events/${slug}.md`, "utf-8");
-  const { data: frontmatter, content } = matter(fileName);
-  return {
-    props: {
-      frontmatter,
-      content,
-    },
-  };
+  return getFileReadStaticProps(slug);
 }
